requirements.txt
flask
python-telegram-bot==20.7
yt-dlp
ffmpeg-python
requests
gunicorn
quart
hypercorn